// src/app/services/api/terraswap_pair/query_msg.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type QueryMsg =
  | {
      pair: {
        [k: string]: unknown;
      };
      [k: string]: unknown;
    }
  | {
      pool: {
        [k: string]: unknown;
      };
      [k: string]: unknown;
    }
  | {
      simulation: {
        offer_asset: Asset;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    }
  | {
      reverse_simulation: {
        ask_asset: Asset;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
export type Uint128 = string;
export type AssetInfo =
  | {
      token: {
        contract_addr: HumanAddr;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    }
  | {
      native_token: {
        denom: string;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
export type HumanAddr = string;

export interface Asset {
  amount: Uint128;
  info: AssetInfo;
  [k: string]: unknown;
}