// src/app/services/api/terraswap_pair/handle_msg.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type HandleMsg =
  | {
      receive: Cw20ReceiveMsg;
      [k: string]: unknown;
    }
  | {
      post_initialize: {
        [k: string]: unknown;
      };
      [k: string]: unknown;
    }
  | {
      provide_liquidity: {
        assets: [Asset, Asset];
        slippage_tolerance?: Decimal | null;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    }
  | {
      swap: {
        belief_price?: Decimal | null;
        max_spread?: Decimal | null;
        offer_asset: Asset;
        to?: HumanAddr | null;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
export type Uint128 = string;
/**
 * Binary is a wrapper around Vec<u8> to add base64 de/serialization with serde. It also adds some helper methods to help encode inline.
 *
 * This is only needed as serde-json-{core,wasm} has a horrible encoding for Vec<u8>
 */
export type Binary = string;
export type HumanAddr = string;
export type AssetInfo =
  | {
      token: {
        contract_addr: HumanAddr;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    }
  | {
      native_token: {
        denom: string;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
/**
 * A fixed-point decimal value with 18 fractional digits, i.e. Decimal(1_000_000_000_000_000_000) == 1.0
 *
 * The greatest possible value that can be represented is 340282366920938463463.374607431768211455 (which is (2^128 - 1) / 10^18)
 */
export type Decimal = string;

/**
 * Cw20ReceiveMsg should be de/serialized under `Receive()` variant in a HandleMsg
 */
export interface Cw20ReceiveMsg {
  amount: Uint128;
  msg?: Binary | null;
  sender: HumanAddr;
  [k: string]: unknown;
}
export interface Asset {
  amount: Uint128;
  info: AssetInfo;
  [k: string]: unknown;
}